+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 Change Log
 ==========
+Version 5.0.0 *(April 30 2026)*
+-------------------------------------------
+**What's new**
+* **[Android Platform]**
+    * Supports [CleverTap Android SDK v8.1.0](https://github.com/CleverTap/clevertap-android-sdk/blob/master/docs/CTCORECHANGELOG.md#version-810-april-17-2026).
+    * Bumps the minimum supported Android API level from 21 (Android 5.0) to 23 (Android 6.0).
+
+* **[iOS Platform]**
+    * Supports [CleverTap iOS SDK v7.6.0](https://github.com/CleverTap/clevertap-ios-sdk/blob/master/CHANGELOG.md#version-760-april-17-2026).
+
+**API changes**
+* **[Android and iOS Platform]**
+    * Adds new `variants(successCallback)` API to fetch the active A/B experiment variants for the current user. Each variant object contains an `"id"` key mapping to the numeric ID of the variant.
+    * Updates `discardInAppNotifications()` with an optional boolean parameter `dismissInAppIfVisible`. When `true`, immediately dismisses any currently visible In-App notification in addition to clearing the pending queue.
+        * `discardInAppNotifications(dismissInAppIfVisible)`
+    * Adds new `unmute()` API to resumes event tracking, overriding any active mute period.
+
+**Bug Fixes**
+* **[iOS Platform]**
+    * Fixes a bug where onUserLogin was creating additional guids only for encryption level high.
+
 Version 4.3.0 *(October 8 2025)*
 -------------------------------------------
 **What's new**